Migrate project to SDL3

+++ src/engine/Timer.h
@@ -1,7 +1,7 @@
 #ifndef __TIMER_H__
 #define __TIMER_H__
 
-#include <SDL.h>
+#include <SDL3/SDL.h>
 
 static const int SECOND = 1000;
 static const int MINUTE = 60 * SECOND;
@@ -13,13 +13,13 @@ static const int HOUR = 60 * MINUTE;
  */
 class Timer {
 	private:
-		Uint32 time;
+		Uint64 time;
 	public:
 		Timer();
 		void measure();
 		void reset();
-		Uint32 getTime();
-		Uint32 getElapsed();
+		Uint64 getTime();
+		Uint64 getElapsed();
 };
 
 #endif
